userSetup.mel: line 42 updated from zooTrigUpdateState; to zooTrigUpdateUIState;
zooTriggeredUtils.mel:
changed zooGetConnects to correctly order the array returned
in zooPopulateCmdStrRaw, regular expression match string changed to %[0-9]+, instead of (%[0-9]+)+... this means strings of the sort "%1%2" will be substituted correctly 

zooObjMenu.mel:
Changed zooObjMenuSaveCmd so it checks the enabled state of the command field before changing
the command field... this way it won't save a populated command if you click close or
change the name while the preview button is checked (but WILL still save the name
change).  Also, I added a save command when a new command is selected to be edited -
either by the right click menu, or when the integer value for the current command is
changed.

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@49 c49f8b0a-e323-0410-bec0-1562a9e7ed1a