p4: bugfix with the return status of edit and revert added a few menu niceties to the p4menu script as well
surgeon:
added a proc to return the list of faces affected by a joint - this is a little more complicated than just converting the vert selection to faces

tangentWks:
parent option added to build menu proc, so the menu can be parented to other parts of the UI

triggerator:
triggers get a right click menu that allows instant editing of the trigger

triggered:
connect name caching - allows triggered to keep working even if a connect gets disconnected but its name remains the same.  this happens for example, when a mesh gets deleted and reimported, but its name remains the same.  commands are available to reconnect disconnected connects.  cache consistency can be an issue, as there is no re-caching on node rename, but this may be integrated into connect querying at a later date

visMan:
hacky implementation of a set merge added - this is pretty nasty, and this tool in particular should probably use a method other than node registration for creating top level sets...

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@72 c49f8b0a-e323-0410-bec0-1562a9e7ed1a