presetManager: did a bunch of p4 integration work.  still needs a bit more for dealing with p4 when renaming/deleting presets in the global area, and adding new preset files...  but the basics are in there.
triggerator:
folder importing now takes an array of input namespaces to search through when performing an import.  ideally this should probably be a user mapping based on namespaces stored in the file, and namespaces preset in the scene, to deal with importing presets that may span multiple namespaces of the same reference file.  but this is a reasonably limited case, so low priority...

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@73 c49f8b0a-e323-0410-bec0-1562a9e7ed1a