quite a few changes here.  most of these are bugfixes or minor feature implementations.
p4 commands now let you add files to a "Maya Auto-checkout" changelist by default - these changes have propagated into a few of the tools.

align and xferAnim now look for the presence of a post trace command attribute to do additional manipulation on the object after an align or a trace has occured.  it defaults to off, but this allows things like transform offsets to be applied to the object after tracing.  useful for tracing rigs that are structurally similar, but have differing orientations

bounds has a new getSkinnedBounds function that returns the bounds of the skinned vert components of a given transform

cam has a toggle mode now so you can toggle through the perspective cameras in the scene easily

greasemonkey now no longer uses referenced monkey nodes

posesToSliders looks for a "controlObject" attribute to determine the default control object now.  this can be setup by rigging scripts so pose sliders are more easily edited

space switching now lets users use either a parent constraint OR an point constraint for when orientation isn't needed.  it could be easily extended to work with orient only constraints as well

fixed the annoying triggered error when selecting components with triggered loaded

plus a heap of cst tidy ups

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@99 c49f8b0a-e323-0410-bec0-1562a9e7ed1a