added a heap of functionality to classes in vectors.py - most of it from various math sites and books around the web.  alot of it hasn't been tested yet, so it may be buggy or just plain wrong.  i'm no math guru
skinWeights.py is a positional based weight save tool - saves weights by both pos and id for restoring later.  is a good companion to zooWeightSave.mel but not quite a replacement.  there is still a bit of extra functionality required to make it superior to weightSave.  but i gotta say - python+pickle=exTREEMe awesome.

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@122 c49f8b0a-e323-0410-bec0-1562a9e7ed1a