looks like I didn't test the interface factory enough.
So I tried to do some refactoring on skeleton builder to use interfaces on the base class.  There is a bunch of overlap between the SkeletonPart and the RigPart classes that looks solvable with interfaces.

But for this to work properly I think I need to add attributes to the interface base class...

git-svn-id: https://zootoolbox.googlecode.com/svn/trunk@293 c49f8b0a-e323-0410-bec0-1562a9e7ed1a