Got lists and tuples working!  No dict support yet though.
Lists and tuples open up a can of worms though - currently SObject checks to see whether the type is serializable when you do the setattr.  But lists make this harder to do.

Currently this check isn't being done on values inside lists/tuples, but it does happen on serialize...  Will revisit this issue later